refactor: 툴팁 관련 리팩토링
