+++ src/components/contentDetail/ScriptList.tsx
@@ -4,9 +4,8 @@ import { ScriptParagraph } from '@/types/interface';
 import { formatTime } from '@/utils/contentDetailUtils';
 import { RefObject } from 'react';
 import ReactPlayer from 'react-player';
-import useHighlightMarks from '@/hooks/contentDetail/useHighlightMarks';
-import useScrappedStore from '@/store/useScrappedStore';
-import { useScrollToElement } from '@/hooks/contentDetail/useScrollToElement';
+import useScriptList from '@/hooks/contentDetail/useScriptList';
+import Tooltip from './Tooltip';
 
 export default function ScriptList({
   fullText,
@@ -17,9 +16,7 @@ export default function ScriptList({
   playerRef: RefObject<ReactPlayer>;
   setPlaying: (playing: boolean) => void;
 }) {
-  const scrappedData = useScrappedStore((state) => state.scrappedData);
-  const { elementRefs } = useScrollToElement();
-  useHighlightMarks(scrappedData, elementRefs);
+  const { tooltip, setTooltip, matched } = useScriptList();
 
   return (
     <div className='md:mb-10 desktop:mb-[37px] desktop:ml-[26px]'>
@@ -37,6 +34,9 @@ export default function ScriptList({
           />
         );
       })}
+      {tooltip.index !== null && matched && (
+        <Tooltip tooltip={tooltip} setTooltip={setTooltip} dictionary={matched} />
+      )}
     </div>
   );
 }
